backend/requirements.txt
# Backend dependencies for Presidio Browser Anonymizer
# Python 3.10+

# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Presidio Core
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354

# NLP
spacy==3.7.2
# Install model: python -m spacy download pl_core_news_md

# Configuration
PyYAML==6.0.1

# Database
sqlite3  # Built-in with Python

# Security
secrets  # Built-in with Python

# Utilities
python-dotenv==1.0.0

# Testing (optional)
pytest==7.4.3
pytest-flask==1.3.0
